Use width converter in transmission
